small fix of soutien-rentree

+++ soutien-rentree-2020-jean-christophe-nguyen-master/src/main/java/fr/univ_amu/iut/exercice1/MoyenneArithmetique.java
@@ -1,11 +1,11 @@
 package fr.univ_amu.iut.exercice1;
 
 public class MoyenneArithmetique {
+
     public double calculer(TableauDeNotes tableau){
         double decimal = 0;
         for(int i=0; i<tableau.size();++i){
-            decimal+=tableau.get(i);
-
+            decimal+=(double) tableau.get(i);
         }
         return decimal;
     }
